Move the 'Add exercise timeline' page
