+++ tests/unit/views/Exercises/Edit/Timeline.spec.js
@@ -1,32 +1,33 @@
-import AddExerciseTimeline from '@/views/AddExerciseTimeline';
+import ExerciseEditTimeline from '@/views/Exercises/Edit/Timeline';
 import { shallowMount, createLocalVue } from '@vue/test-utils';
 import Router from 'vue-router';
 
 const localVue = createLocalVue();
 localVue.use(Router);
 
-describe('AddExerciseTimeline', () => {
+describe('ExerciseEditTimeline', () => {
   it('renders the view', () => {
-    let wrapper = shallowMount(AddExerciseTimeline, {
+    let wrapper = shallowMount(ExerciseEditTimeline, {
       localVue,
     });
 
     expect(wrapper.exists()).toBe(true);
   });
+
   it('contains the link to About The Vacancy page', () => {
-    let wrapper = shallowMount(AddExerciseTimeline, {
+    let wrapper = shallowMount(ExerciseEditTimeline, {
       localVue,
     });
     expect(wrapper.find({ ref: 'linkToAboutTheVacancy' }).isVisible()).toBe(true);
   });
-});
 
-describe('Accessibility:', () => {
-  it('page contains h1 element', () => {
-    let wrapper = shallowMount(AddExerciseTimeline, {
-      localVue,
+  describe('Accessibility:', () => {
+    it('page contains h1 element', () => {
+      let wrapper = shallowMount(ExerciseEditTimeline, {
+        localVue,
+      });
+
+      expect(wrapper.contains('h1')).toBe(true);
     });
-
-    expect(wrapper.contains('h1')).toBe(true);
   });
 });
